tegra: therm_est: make coeffs of type s32
Because we need to read some values from device tree, fixing the type
of data we read to 32-bit long, rather than 'long' that has arch-dependent
size.

Long might be 32- or 64-bit long, so better to be always on the safe side

Bug 1844095

Reviewed-on: http://git-master/r/1260765
GVS: Gerrit_Virtual_Submit
Tested-by: Wahed Syed <wsyed@nvidia.com>
Reviewed-on: http://git-master/r/1270219

+++ include/linux/therm_est.h
@@ -35,7 +35,7 @@ struct therm_est_sub_thz {
 
 struct therm_est_coeffs {
 	long toffset;
-	long (*coeffs)[HIST_LEN];
+	s32 (*coeffs)[HIST_LEN];
 };
 
 struct therm_est_subdevice {
@@ -100,7 +100,13 @@ struct therm_fan_est_subdevice {
 	const char *dev_data;
 	int (*get_temp)(const char *, long *);
 	int group;
-	long coeffs[HIST_LEN];
+	/*
+	 * as we read coeffs array from the device tree,
+	 * specify type that has known width - 32 bits
+	 *
+	 * 'long' does not work, as its size if arch-dependent
+	 */
+	s32 coeffs[HIST_LEN];
 	int hist[HIST_LEN];
 };
